video-generator/requirements.txt
# Streamlit Framework
streamlit>=1.30.0

# Google Cloud Services
google-cloud-texttospeech>=2.16.0
google-genai>=0.3.0
google-auth>=2.25.0

# BGM Generation
# beatoven.ai はREST APIを直接使用（公式SDKなし）

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Document Parsing
python-docx>=1.1.0

# Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
